[EasyUtils] Reorganise classes
