+++ packages/EasyUtils/src/MathComparison.php
@@ -4,53 +4,11 @@ declare(strict_types=1);
 
 namespace EonX\EasyUtils;
 
-use EonX\EasyUtils\Interfaces\MathComparisonInterface;
+use EonX\EasyUtils\Math\MathComparison as BaseMathComparison;
 
-final class MathComparison implements MathComparisonInterface
+/**
+ * @deprecated since 4.0, will be removed in 5.0. Use EonX\EasyUtils\Math\MathComparison instead.
+ */
+final class MathComparison extends BaseMathComparison
 {
-    /**
-     * @var string
-     */
-    private $leftOperand;
-
-    /**
-     * @var int
-     */
-    private $scale;
-
-    public function __construct(string $leftOperand, int $scale)
-    {
-        $this->leftOperand = $leftOperand;
-        $this->scale = $scale;
-    }
-
-    public function equalTo(string $rightOperand): bool
-    {
-        return $this->comp($this->leftOperand, $rightOperand) === 0;
-    }
-
-    public function greaterOrEqualTo(string $rightOperand): bool
-    {
-        return $this->comp($this->leftOperand, $rightOperand) !== -1;
-    }
-
-    public function greaterThan(string $rightOperand): bool
-    {
-        return $this->comp($this->leftOperand, $rightOperand) === 1;
-    }
-
-    public function lessOrEqualTo(string $rightOperand): bool
-    {
-        return $this->comp($this->leftOperand, $rightOperand) !== 1;
-    }
-
-    public function lessThan(string $rightOperand): bool
-    {
-        return $this->comp($this->leftOperand, $rightOperand) === -1;
-    }
-
-    private function comp(string $leftOperand, string $rightOperand): int
-    {
-        return \bccomp($leftOperand, $rightOperand, $this->scale);
-    }
 }
